requirements.txt
absl-py==0.13.0
anyio==2.2.0
APScheduler==3.6.3
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
caer==1.9.8
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
comtypes==1.1.10
covid==2.4.0
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
distlib==0.3.2
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
gym==0.18.3
h5py==3.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.71
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlip==0.2
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mypy==0.910
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.0
pipwin==0.5.1
platformdirs==2.2.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Pavana%20H%20A/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycparser==2.20
pydantic==1.8.1
pyfiglet==0.8.post1
PyGetWindow==0.0.9
pyglet==1.5.15
Pygments==2.8.1
pyjsparser==2.7.1
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pySmartDL==1.3.4
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.9
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.1
SpeechRecognition==3.8.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
wikipedia==1.4.0
wrapt==1.12.1
